백준-ACM Hotel 문제풀이

+++ CodingTestOnJava/src/programmers/DartGame.java
@@ -30,7 +30,7 @@ public class DartGame {
         	}
         	
         	if(dartResult.charAt(i)=='D') {
-        		score[++idx] = (int)(Math.pow((double)num, 2.0));
+        		score[++idx] = (int) Math.pow(num, 2.0);
         		num = 0;
         		continue;
         	}
